refactor(best-practices): `BestPractices` is now `BestPracticesRunner`, `BestPracticesModule` is now `BestPractices`
This should better align meaning with readability of external API.
